Added deploy script
Script to export docker images as tar and send it to the server. Once
on the server, it's loaded and then the container is stopped and
started via docker-compose.

This will make updates much more regular since deployment went from a
manual process taking upwards of 15 minutes, to an automated process
taking 30 seconds.